Restructure backend into modular package
