+++ backend/app/config/redis.py
@@ -1,0 +1,2 @@
+# Placeholder for Redis configuration
+redis_client = None
