+++ backend/app/models/user.py
@@ -1,6 +1,4 @@
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
+from ..config.database import db
 
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
